core - dynamic ports impl

+++ src/core/ipc/pipe.h
@@ -11,6 +11,6 @@ extern "C"
 			FRONT_END_LOADED
 		};
 
-		const int OpenConnection();
+		const uint16_t OpenConnection();
 	}
 }
